Beta 0.5.5
Fixed a bug where all tasks would be added to the same core. Added the
ability for the player to "mine" a block. It is set to deal a cetain
amount of "damage" every wait period. There is no visual representation
as to when the player is mining. It will start when you click and end
when you release the mouse button. The block will remember its health
inbetween clicks.