Beta 0.6.4
Fixed all bugs with the merger of the two versions, except the player
possition one being kinda wonky. Also, added the ability to save your
inventory. 

For Devs: Changed the way it prints out time taken to do something, now
it will only print the time taken to do that method, not the total
ellapsed time.